separate cost utils from runtime utils in new cost-utils.js; move style and component helpers from runtime-utils into runtime-common
