+++ src/data/gce-machines.js
