+++ src/pages/workspaces/workspace/analysis/RuntimeManager.js
@@ -6,7 +6,7 @@ import { Fragment, useEffect, useState } from 'react'
 import { div, h, p } from 'react-hyperscript-helpers'
 import { ButtonPrimary, Clickable, Link, spinnerOverlay } from 'src/components/common'
 import Modal from 'src/components/Modal'
-import { dataSyncingDocUrl } from 'src/data/machines'
+import { dataSyncingDocUrl } from 'src/data/gce-machines'
 import { Ajax } from 'src/libs/ajax'
 import { getDynamic, getSessionStorage, setDynamic } from 'src/libs/browser-storage'
 import colors from 'src/libs/colors'
