+++ src/pages/workspaces/workspace/analysis/runtime-utils.js
@@ -1,25 +1,10 @@
 import _ from 'lodash/fp'
-import { Fragment } from 'react'
-import { div, h, input, label } from 'react-hyperscript-helpers'
-import { IdContainer } from 'src/components/common'
 import {
-  cloudServices, dataprocCpuPrice, ephemeralExternalIpAddressPrice, gpuTypes, machineTypes, regionToPrices, zonesToGpus
-} from 'src/data/machines'
-import colors from 'src/libs/colors'
-import * as Style from 'src/libs/style'
+  cloudServices, gpuTypes, machineTypes, zonesToGpus
+} from 'src/data/gce-machines'
 import * as Utils from 'src/libs/utils'
 import { allAppTypes, appTools, toolLabels } from 'src/pages/workspaces/workspace/analysis/tool-utils'
 
-
-export const computeStyles = {
-  label: { fontWeight: 600, whiteSpace: 'pre' },
-  value: { fontWeight: 400, whiteSpace: 'pre' },
-  titleBar: { marginBottom: '1rem' },
-  whiteBoxContainer: { padding: '1rem', borderRadius: 3, backgroundColor: 'white' },
-  drawerContent: { display: 'flex', flexDirection: 'column', flex: 1, padding: '1.5rem' },
-  headerText: { fontSize: 16, fontWeight: 600 },
-  warningView: { backgroundColor: colors.warning(0.1) }
-}
 
 export const pdTypes = {
   standard: {
@@ -132,151 +117,7 @@ export const getValidGpuOptions = (numCpus, mem, zone) => {
   return validGpuOptions || { name: '?', type: '?', numGpus: '?', maxNumCpus: '?', maxMem: '?', price: NaN, preemptiblePrice: NaN }
 }
 
-const dataprocCost = (machineType, numInstances) => {
-  const { cpu: cpuPrice } = findMachineType(machineType)
-
-  return cpuPrice * numInstances * dataprocCpuPrice
-}
-
-const getHourlyCostForMachineType = (machineTypeName, region, isPreemptible) => {
-  const { cpu, memory } = _.find({ name: machineTypeName }, machineTypes) || {}
-  const { n1HourlyCpuPrice, preemptibleN1HourlyCpuPrice, n1HourlyGBRamPrice, preemptibleN1HourlyGBRamPrice } = _.find({ name: _.toUpper(region) },
-    regionToPrices) || {}
-  return isPreemptible ?
-    (cpu * preemptibleN1HourlyCpuPrice) + (memory * preemptibleN1HourlyGBRamPrice) :
-    (cpu * n1HourlyCpuPrice) + (memory * n1HourlyGBRamPrice)
-}
-
-const getGpuCost = (gpuType, numGpus, region) => {
-  const prices = _.find({ name: region }, regionToPrices)
-  // From a type like 'nvidia-tesla-t4', look up 't4HourlyPrice' in prices
-  const price = prices[`${_.last(_.split('-', gpuType))}HourlyPrice`]
-  return price * numGpus
-}
-
-export const runtimeConfigBaseCost = config => {
-  const {
-    cloudService, masterMachineType, masterDiskSize, numberOfWorkers, workerMachineType, workerDiskSize, bootDiskSize, computeRegion
-  } = normalizeRuntimeConfig(config)
-
-  const isDataproc = cloudService === cloudServices.DATAPROC
-
-  return _.sum([
-    (masterDiskSize + numberOfWorkers * workerDiskSize) * getPersistentDiskPriceForRegionHourly(computeRegion, pdTypes.standard),
-    isDataproc ?
-      (dataprocCost(masterMachineType, 1) + dataprocCost(workerMachineType, numberOfWorkers)) :
-      (bootDiskSize * getPersistentDiskPriceForRegionHourly(computeRegion, pdTypes.standard))
-  ])
-}
-
-export const runtimeConfigCost = config => {
-  const {
-    cloudService, masterMachineType, numberOfWorkers, numberOfPreemptibleWorkers, workerMachineType, workerDiskSize, computeRegion
-  } = normalizeRuntimeConfig(
-    config)
-  const masterPrice = getHourlyCostForMachineType(masterMachineType, computeRegion, false)
-  const workerPrice = getHourlyCostForMachineType(workerMachineType, computeRegion, false)
-  const preemptiblePrice = getHourlyCostForMachineType(workerMachineType, computeRegion, true)
-  const numberOfStandardVms = 1 + numberOfWorkers // 1 is for the master VM
-  const gpuConfig = config?.gpuConfig
-  const gpuEnabled = cloudService === cloudServices.GCE && !!gpuConfig
-
-  return _.sum([
-    masterPrice,
-    numberOfWorkers * workerPrice,
-    numberOfPreemptibleWorkers * preemptiblePrice,
-    numberOfPreemptibleWorkers * workerDiskSize * getPersistentDiskPriceForRegionHourly(computeRegion, pdTypes.standard),
-    cloudService === cloudServices.DATAPROC && dataprocCost(workerMachineType, numberOfPreemptibleWorkers),
-    gpuEnabled && getGpuCost(gpuConfig.gpuType, gpuConfig.numOfGpus, computeRegion),
-    ephemeralExternalIpAddressCost({ numStandardVms: numberOfStandardVms, numPreemptibleVms: numberOfPreemptibleWorkers }),
-    runtimeConfigBaseCost(config)
-  ])
-}
-
-// Per GB following https://cloud.google.com/compute/pricing
-const getPersistentDiskPriceForRegionMonthly = (computeRegion, diskType) => {
-  return _.flow(_.find({ name: _.toUpper(computeRegion) }), _.get([diskType.regionToPricesName]))(regionToPrices)
-}
-const numberOfHoursPerMonth = 730
-const getPersistentDiskPriceForRegionHourly = (computeRegion, diskType) => getPersistentDiskPriceForRegionMonthly(computeRegion, diskType) / numberOfHoursPerMonth
-
-export const getPersistentDiskCostMonthly = (currentPersistentDiskDetails, computeRegion) => {
-  const price = getPersistentDiskPriceForRegionMonthly(computeRegion, currentPersistentDiskDetails?.diskType)
-  return _.includes(currentPersistentDiskDetails?.status, ['Deleting', 'Failed']) ? 0.0 : currentPersistentDiskDetails?.size * price
-}
-export const getPersistentDiskCostHourly = ({ size, status, diskType }, computeRegion) => {
-  const price = getPersistentDiskPriceForRegionHourly(computeRegion, diskType)
-  return _.includes(status, ['Deleting', 'Failed']) ? 0.0 : size * price
-}
-
-const ephemeralExternalIpAddressCost = ({ numStandardVms, numPreemptibleVms }) => {
-  // Google categorizes a VM as 'standard' if it is not 'pre-emptible'.
-  return numStandardVms * ephemeralExternalIpAddressPrice.standard + numPreemptibleVms * ephemeralExternalIpAddressPrice.preemptible
-}
-
-export const getRuntimeCost = ({ runtimeConfig, status }) => Utils.switchCase(status,
-  [
-    'Stopped',
-    () => runtimeConfigBaseCost(runtimeConfig)
-  ],
-  ['Error', () => 0.0],
-  [Utils.DEFAULT, () => runtimeConfigCost(runtimeConfig)]
-)
-
-
 export const isApp = cloudEnvironment => !!cloudEnvironment?.appName
-
-export const getAppCost = (app, dataDisk) => app.appType === appTools.Galaxy.appType ? getGalaxyCost(app, dataDisk) : 0
-
-export const getGalaxyCost = (app, dataDisk) => {
-  return getGalaxyDiskCost(dataDisk) + getGalaxyComputeCost(app)
-}
-
-/*
- * - Default nodepool VMs always run and incur compute and external IP cost whereas app
- *   nodepool VMs incur compute and external IP cost only when an app is running.
- * - Default nodepool cost is shared across all Kubernetes cluster users. It would
- *   be complicated to calculate that shared cost dynamically. Therefore, we are being
- *   conservative by adding default nodepool cost to all apps on a cluster.
- */
-export const getGalaxyComputeCost = app => {
-  const appStatus = app?.status?.toUpperCase()
-  // Galaxy uses defaultComputeRegion because we're not yet enabling other locations for Galaxy apps.
-  const defaultNodepoolComputeCost = getHourlyCostForMachineType(defaultGceMachineType, defaultComputeRegion, false)
-  const defaultNodepoolIpAddressCost = ephemeralExternalIpAddressCost({ numStandardVms: 1, numPreemptibleVms: 0 })
-
-  const staticCost = defaultNodepoolComputeCost + defaultNodepoolIpAddressCost
-  const dynamicCost = app.kubernetesRuntimeConfig.numNodes *
-    getHourlyCostForMachineType(app.kubernetesRuntimeConfig.machineType, defaultComputeRegion, false) +
-    ephemeralExternalIpAddressCost({ numStandardVms: app.kubernetesRuntimeConfig.numNodes, numPreemptibleVms: 0 })
-
-  switch (appStatus) {
-    case 'STOPPED':
-      return staticCost
-    case 'DELETING':
-    case 'ERROR':
-      return 0.0
-    default:
-      return staticCost + dynamicCost
-  }
-}
-
-/*
- * - Disk cost is incurred regardless of app status.
- * - Disk cost is total for data (NFS) disk, metadata (postgres) disk, and boot disks (1 boot disk per nodepool)
- * - Size of a data disk is user-customizable. The other disks have fixed sizes.
- */
-export const getGalaxyDiskCost = ({ size: dataDiskType, diskType }) => {
-  const metadataDiskSize = 10 // GB
-  const defaultNodepoolBootDiskSize = 100 // GB
-  const appNodepoolBootDiskSize = 100 // GB
-
-  return getPersistentDiskCostHourly({
-    status: 'Running',
-    size: dataDiskType + metadataDiskSize + defaultNodepoolBootDiskSize + appNodepoolBootDiskSize,
-    diskType
-  }, defaultComputeRegion)
-}
 
 export const trimRuntimesOldestFirst = _.flow(
   _.remove({ status: 'Deleting' }),
@@ -376,44 +217,6 @@ export const isCurrentGalaxyDiskDetaching = apps => {
   return currentGalaxyApp && _.includes(currentGalaxyApp.status, ['DELETING', 'PREDELETING'])
 }
 
-export const getGalaxyCostTextChildren = (app, appDataDisks) => {
-  const dataDisk = getCurrentAttachedDataDisk(app, appDataDisks)
-  return app ?
-    [getComputeStatusForDisplay(app.status), dataDisk ? ` ${Utils.formatUSD(getGalaxyCost(app, dataDisk))}/hr` : ''] : ['']
-}
-
-// TODO: multiple runtime: this is a good example of how the code should look when multiple runtimes are allowed, over a tool-centric approach
-export const getCostDisplayForTool = (app, currentRuntime, currentRuntimeTool, toolLabel) => {
-  return Utils.cond(
-    [toolLabel === toolLabels.Galaxy, () => app ? `${getComputeStatusForDisplay(app.status)} ${Utils.formatUSD(getGalaxyComputeCost(app))}/hr` : ''],
-    [toolLabel === toolLabels.Cromwell, () => ''], // We will determine what to put here later
-    [toolLabel === toolLabels.JupyterLab, () => ''], //TODO: Azure cost calculation
-    [getRuntimeForTool(toolLabel, currentRuntime, currentRuntimeTool), () => `${getComputeStatusForDisplay(currentRuntime.status)} ${Utils.formatUSD(getRuntimeCost(currentRuntime))}/hr`],
-    [Utils.DEFAULT, () => {
-      return ''
-    }]
-  )
-}
-
-export const getCostDisplayForDisk = (app, appDataDisks, computeRegion, currentRuntimeTool, persistentDisks, runtimes, toolLabel) => {
-  const diskCost = getCostForDisk(app, appDataDisks, computeRegion, currentRuntimeTool, persistentDisks, runtimes, toolLabel)
-  return diskCost ? `Disk ${Utils.formatUSD(diskCost)}/hr` : ''
-}
-
-export const getCostForDisk = (app, appDataDisks, computeRegion, currentRuntimeTool, persistentDisks, runtimes, toolLabel) => {
-  let diskCost = ''
-  if (currentRuntimeTool === toolLabel && persistentDisks && persistentDisks.length > 0) {
-    const curPd = getCurrentPersistentDisk(runtimes, persistentDisks)
-    const { size = 0, status = 'Running', diskType = pdTypes.standard } = curPd || {}
-    diskCost = getPersistentDiskCostHourly({ size, status, diskType }, computeRegion)
-  } else if (app && appDataDisks && (toolLabel === 'Galaxy')) {
-    const currentDataDisk = getCurrentAttachedDataDisk(app, appDataDisks)
-    //Occasionally currentDataDisk will be undefined on initial render.
-    diskCost = currentDataDisk ? getGalaxyDiskCost(currentDataDisk) : ''
-  }
-  return diskCost
-}
-
 
 // TODO: multiple runtime: build component around this logic for a multiple runtime approach. see getCostForTool for example usage
 export const getRuntimeForTool = (toolLabel, currentRuntime, currentRuntimeTool) => Utils.cond([toolLabel === currentRuntimeTool, () => currentRuntime],
@@ -470,26 +273,6 @@ export const displayNameForGpuType = type => {
     ['nvidia-tesla-p100', () => 'NVIDIA Tesla P100'],
     [Utils.DEFAULT, () => 'NVIDIA Tesla T4']
   )
-}
-
-export const RadioBlock = ({ labelText, children, name, checked, onChange, style = {} }) => {
-  return div({
-    style: {
-      backgroundColor: colors.warning(0.2),
-      borderRadius: 3, border: `1px solid ${checked ? colors.accent() : 'transparent'}`,
-      boxShadow: checked ? Style.standardShadow : undefined,
-      display: 'flex', alignItems: 'baseline', padding: '.75rem',
-      ...style
-    }
-  }, [
-    h(IdContainer, [id => h(Fragment, [
-      input({ type: 'radio', name, checked, onChange, id }),
-      div({ style: { marginLeft: '.75rem' } }, [
-        label({ style: { fontWeight: 600, fontSize: 16 }, htmlFor: id }, [labelText]),
-        children
-      ])
-    ])])
-  ])
 }
 
 export const getIsAppBusy = app => app?.status !== 'RUNNING' && _.includes('ING', app?.status)
